修改“替换curl版本”的判断条件

+++ immortalwrt/diy-part2.sh
@@ -105,11 +105,11 @@ sed -i 's/procd_set_param stderr 1/procd_set_param stderr 0/g' feeds/packages/ut
 
 # 替换curl修改版（无nghttp3、ngtcp2）
 curl_ver=$(cat feeds/packages/net/curl/Makefile | grep -i "PKG_VERSION:=" | awk 'BEGIN{FS="="};{print $2}')
-[ "$(check_ver "$curl_ver" "8.12.0")" != "0" ] && {
+if [ "$(check_ver "$curl_ver" "8.12.0")" != "0" ] || [ "$curl_ver" = "8.12.1" ]; then
 	echo "替换curl版本"
 	rm -rf feeds/packages/net/curl
 	cp -rf ${GITHUB_WORKSPACE}/patch/curl feeds/packages/net/curl
-}
+fi
 
 # apk-tools APK管理器不再校验版本号的合法性
 mkdir -p package/system/apk/patches && cp -f ${GITHUB_WORKSPACE}/patch/apk-tools/9999-hack-for-linux-pre-releases.patch package/system/apk/patches/
